refactor: split to path
