+++ src/handler/basic/cat.js
@@ -1,8 +1,8 @@
 import { createReadStream, constants } from 'fs';
 import { resolve } from 'path';
-import { dispCurrentDir } from '../helper/index.js';
+import { dispCurrentDir } from '../../helper/index.js';
 
-export const readFile = async (fileName) => {
+export const catHandler = async (fileName) => {
   try {
     const currentPath = resolve(process.cwd());
     const filePath = resolve(currentPath, String(fileName));
