+++ src/handler/index.js
@@ -1,10 +1,13 @@
-export * from './ls.js';
-export * from './cd.js';
-export * from './up.js';
-export * from './cat.js';
-export * from './add.js';
-export * from './rm.js';
-export * from './rn.js';
-export * from './cp.js';
-export * from './mv.js';
-export * from './os/index.js';
+import * as basic from './basic/index.js';
+
+export const lsHandler = basic.lsHandler;
+export const cdHandler = basic.cdHandler;
+export const upHandler = basic.upHandler;
+export const catHandler = basic.catHandler;
+export const addHandler = basic.addHandler;
+export const rmHandler = basic.rmHandler;
+export const rnHandler = basic.rnHandler;
+export const cpHandler = basic.cpHandler;
+export const mvHandler = basic.mvHandler;
+
+
